+++ src/rip/operations/DistanceTableParser.java
@@ -3,13 +3,13 @@ package rip.operations;
 import java.util.List;
 
 public class DistanceTableParser {
-    public static short[][] parse(String data) {
+    public static int[][] parse(String data) {
         List<String> strList = List.of(data.split(" "));
-        short[][] distanceTable = new short[strList.size()][];
+        int[][] distanceTable = new int[strList.size()][];
         try {
             for (int i = 0; i < strList.size(); i++) {
                 String str = strList.get(i);
-                short[] vec = DistanceVectorParser.parse(str);
+                int[] vec = DistanceVectorParser.parse(str);
                 if (vec == null) {
                     return null;
                 }
